test(background-sync): absorb sync-and-utils-boost, delete the padding file
Decouple tests/data/sync-and-utils-boost.test.ts (audit finding F2.1). Its
task-mapper, quadrants, id-generator, utils, and routes blocks duplicated the
canonical module tests (which are more thorough) and were dropped. Of the 13
BackgroundSyncManager cases, 9 duplicated background-sync.test.ts; the 4 that
covered unique branches (syncOnFocus/syncOnOnline listeners, the
scheduleDebouncedSync active path, and restart-stops-previous) were migrated
into the canonical file. The debounce case was rewritten to assert the sync
actually fires (the original asserted nothing) — starting with enabled:false so
the debounced sync clears the min-interval guard.

Verified: full suite green (1942 pass); total coverage up slightly
(lines 86.74%->86.76%); all six affected modules hold or improve their
coverage (background-sync functions 78.6%->85.7%). One padding file removed.
